Remove experimental features
